+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 # Changelog
 
+## 1.0.0
+
+- [x] CRUD endpoints
+- [x] Unit tests
+- [x] Acceptance tests
+- [x] Utils
+- [x] Drone pipeline
+
 ## 0.1.0
 
 - [x] Minimal functionality / endpoint with simple hello